Move getACL to a utility class
